Fix repair cost for steel inheriting from iron (A19/ULM)
If wasn't possible to simply set the repair cost to zero,
as the game still would require at least on of that item.
Found this fancy `param1` attribute for `Extends`, which
seems to actually do what I need. I guess it just tells
it to ignore that property while inheriting/extending!?